Moved provider test to provider_test.go
